--- app/views/admin/pages/tabs/_snippets.html.erb
@@ -1,3 +1,0 @@
-<div class="wym_skin_refinery page_part" id="page_snippet_picker">
-  <%= render :partial => '/admin/pages/tabs/snippets_content', :locals => {:f => f} %>
-</div>
