--- app/views/admin/snippets_page_parts/remove.html.erb
@@ -1,1 +1,0 @@
-<%= render :partial => '/admin/pages/tabs/snippets_content' %>
