+++ app/views/refinery/admin/pages/tabs/_snippets.html.erb
@@ -1,0 +1,3 @@
+<div class="wym_skin_refinery page_part" id="page_snippet_picker">
+  <%= render :partial => '/refinery/admin/pages/tabs/snippets_content', :locals => {:f => f} %>
+</div>
