+++ app/views/refinery/admin/pages/tabs/_snippets_list_item.html.erb
@@ -11,9 +11,9 @@
     <% end %>
   </span>
   <span class='actions'>
-    <%= link_to refinery_icon_tag('application_edit.png'), edit_admin_snippet_path(snippet),
+    <%= link_to refinery_icon_tag('application_edit.png'), main_app.edit_refinery_admin_snippet_path(snippet),
         :title => t('.edit') %>
-    <%= link_to refinery_icon_tag('delete.png'), url_for({:controller => 'snippets_page_parts', :action => 'remove', :id => @page.id, :snippet_id => snippet.id, :part_id => part.id, :before_body => before_body}),
+    <%= link_to refinery_icon_tag('delete.png'), main_app.url_for({:controller => 'snippets_page_parts', :action => 'remove', :id => @page.id, :snippet_id => snippet.id, :part_id => part.id, :before_body => before_body}),
         :title => t('.remove'),
         :class => 'remove-snippet' %>
   </span>
