+++ app/views/refinery/admin/snippets/_actions.html.erb
@@ -1,25 +1,25 @@
 <ul>
-  <% if Admin::SnippetsController.searchable? %>
+  <% if Refinery::Admin::SnippetsController.searchable? %>
     <li>
-      <%= render :partial => "/shared/admin/search",
+      <%= render :partial => "/refinery/admin/search",
                   :locals => {
-                    :url => admin_snippets_url
+                    :url => main_app.refinery_admin_snippets_path
                   } %>
     </li>
   <% end %>
   <li>
-    <%= link_to t('.create_new'), new_admin_snippet_url,
+    <%= link_to t('.create_new'), main_app.new_refinery_admin_snippet_path,
                  :class => "add_icon" %>
   </li>
-<% if !searching? and Admin::SnippetsController.sortable? and Snippet.count > 1 %>
+<% if !searching? and Refinery::Admin::SnippetsController.sortable? and Refinery::Snippet.count > 1 %>
   <li>
     <%= link_to t('.reorder', :what => "Snippets"),
-                 admin_snippets_url,
+                 main_app.refinery_admin_snippets_path,
                  :id => "reorder_action",
                  :class => "reorder_icon" %>
 
     <%= link_to t('.reorder_done', :what => "Snippets"),
-                 admin_snippets_url,
+                 main_app.refinery_admin_snippets_path,
                  :id => "reorder_action_done",
                  :style => "display: none;",
                  :class => "reorder_icon" %>
