+++ app/views/refinery/admin/snippets/_form.html.erb
@@ -1,10 +1,9 @@
-<%= form_for [:admin, @snippet] do |f| -%>
-  <%= render :partial => "/shared/admin/error_messages", :locals => {
+<%= form_for [main_app, :refinery_admin, @snippet] do |f| -%>
+  <%= render :partial => "/refinery/admin/error_messages", :locals => {
           :object => @snippet,
           :include_object_name => true
         } %>
   
-
   <%= render :partial => "locale_picker",
              :locals => {
                :current_locale => Thread.current[:globalize_locale]
@@ -34,6 +33,3 @@
                 :delete_confirmation => t('message', :scope => 'shared.admin.delete', :title => @snippet.title)
               } %>
 <% end -%>
-<% content_for :javascripts do %>
-
-<% end %>
