+++ app/views/refinery/admin/snippets/_records.html.erb
