+++ app/views/refinery/admin/snippets/_snippet.html.erb
@@ -11,9 +11,9 @@
     <% end %>
   </span>
   <span class='actions'>
-    <%= link_to refinery_icon_tag("application_edit.png"), edit_admin_snippet_path(snippet),
+    <%= link_to refinery_icon_tag("application_edit.png"), main_app.edit_refinery_admin_snippet_path(snippet),
          :title => t('.edit') %>
-    <%= link_to refinery_icon_tag("delete.png"), admin_snippet_path(snippet),
+    <%= link_to refinery_icon_tag("delete.png"), main_app.refinery_admin_snippet_path(snippet),
         :class => "cancel confirm-delete",
         :title => t('.delete'),
         :confirm => t('message', :scope => 'shared.admin.delete', :title => snippet.title),
