+++ app/views/refinery/admin/snippets/_sortable_list.html.erb
@@ -1,7 +1,7 @@
 <ul id='sortable_list'>
   <%= render :partial => 'snippet', :collection => @snippets %>
 </ul>
-<%= render :partial => "/shared/admin/sortable_list",
+<%= render :partial => "/refinery/admin/sortable_list",
             :locals => {
               :continue_reordering => (local_assigns.keys.include?(:continue_reordering)) ? continue_reordering : true
             } %>
