+++ app/views/refinery/admin/snippets/new.html.erb
