+++ app/views/refinery/admin/snippets_page_parts/remove.html.erb
@@ -1,0 +1,1 @@
+<%= render :partial => '/refinery/admin/pages/tabs/snippets_content' %>
