+++ db/migrate/1_create_snippets.rb
@@ -1,22 +1,22 @@
 class CreateSnippets < ActiveRecord::Migration
 
   def self.up
-    create_table :snippets do |t|
-      t.string :title, :limit => 36, :null => false
-      t.text :body
-      t.integer :position, :null => false, :default => 0
-      t.timestamps
+    unless Refinery::Snippet.table_exists?
+      create_table Refinery::Snippet.table_name do |t|
+        t.string :title, :limit => 36, :null => false
+        t.text :body
+        t.integer :position, :null => false, :default => 0
+        t.timestamps
+      end
     end
-
+    
     load(Rails.root.join('db', 'seeds', 'snippets.rb'))
   end
 
   def self.down
-    UserPlugin.destroy_all({:name => "snippets"})
+    Refinery::UserPlugin.destroy_all({:name => "snippets"})
 
-    Page.delete_all({:link_url => "/snippets"})
-
-    drop_table :snippets
+    drop_table Refinery::Snippet.table_name
   end
 
 end
