+++ lib/generators/refinery/refinerycms_snippets_generator.rb
@@ -1,0 +1,8 @@
+module ::Refinery
+  class RefinerycmsSnippetsGenerator < ::Refinery::Generators::EngineInstaller
+
+    source_root File.expand_path('../../../../', __FILE__)
+    engine_name "snippets"
+
+  end
+end
