--- lib/generators/refinerycms_snippets_generator.rb
@@ -1,6 +1,0 @@
-class RefinerycmsSnippets < Refinery::Generators::EngineInstaller
-
-  source_root File.expand_path('../../../', __FILE__)
-  engine_name "snippets"
-
-end
