+++ lib/refinerycms-snippets.rb
@@ -2,6 +2,14 @@ require 'refinerycms-base'
 
 module Refinery
   module Snippets
+    
+    class << self
+      attr_accessor :root
+      def root
+        @root ||= Pathname.new(File.expand_path('../../', __FILE__))
+      end
+    end
+    
     class Engine < Rails::Engine
 
       config.before_initialize do
@@ -14,28 +22,30 @@ module Refinery
       end
 
       config.to_prepare do
-        PagePart.module_eval do
+        Refinery::PagePart.module_eval do
           has_many :snippet_page_parts, :dependent => :destroy
           has_many :snippets, :through => :snippet_page_parts, :order => 'position ASC'
         end
-        Page.send :include, Extensions::Page
-        PagesHelper.send :include, Extensions::PagesHelper
+        Refinery::Page.send :include, Extensions::Page
+        Refinery::PagesHelper.send :include, Extensions::PagesHelper
       end
 
+      
       config.after_initialize do
-        ::Refinery::Pages::Tab.register do |tab|
-          tab.name = "snippets"
-          tab.partial = "/admin/pages/tabs/snippets"
-        end
         Refinery::Plugin.register do |plugin|
+          plugin.pathname = root
           plugin.name = "snippets"
-          plugin.url = {:controller => '/admin/snippets'}
-          plugin.menu_match = /^\/?(admin|refinery)\/snippets/
+          plugin.url = '/refinery/snippets'
           plugin.activity = [{
-                               :class => Snippet
+                               :class => Refinery::Snippet
                              }, {
-                               :class => SnippetPage
+                               :class => Refinery::SnippetPage
                              }]
+        end
+        
+        Refinery::Pages::Tab.register do |tab|
+          tab.name = "snippets"
+          tab.partial = "/refinery/admin/pages/tabs/snippets"
         end
       end
     end
